refactor: update Playwright test command for API project and remove obsolete assessment, authentication, and user endpoint tests to streamline test suite
