+++ backend/dev/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "End-to-end API tests for Dottie application",
   "scripts": {
-    "test": "npx playwright test",
+    "test": "npx playwright test --project=api",
     "test:basic": "npx playwright test tests/api-basic-endpoints.api.pw.spec.js",
     "test:auth": "npx playwright test tests/api-auth-endpoints.api.pw.spec.js",
     "test:assessment": "npx playwright test tests/api-assessment-endpoints.api.pw.spec.js",
